ref(bufferline): partial extract `render` into `api`
There was a specific type of out-of-scope function in `render`—
high-level functions that are called by EX-commands. I extracted
these into a new module (`api`) to reduce the noise in `render`.
